+++ config/engine.conf/src/test/java/io/aklivity/zilla/config/engine/test/internal/model/config/TestModelConfigAdapter.java
@@ -19,6 +19,7 @@ import java.util.List;
 
 import jakarta.json.Json;
 import jakarta.json.JsonArray;
+import jakarta.json.JsonNumber;
 import jakarta.json.JsonObject;
 import jakarta.json.JsonString;
 import jakarta.json.JsonValue;
@@ -36,6 +37,7 @@ public class TestModelConfigAdapter extends ConfigAdapter<ModelConfig, JsonValue
     private static final String TEST = "test";
     private static final String LENGTH = "length";
     private static final String TRANSFORM = "transform";
+    private static final String TRANSFORM_AUTHORIZATIONS = "transformAuthorizations";
     private static final String CAPABILITY = "capability";
     private static final String READ = "read";
     private static final String CATALOG_NAME = "catalog";
@@ -62,8 +64,18 @@ public class TestModelConfigAdapter extends ConfigAdapter<ModelConfig, JsonValue
             : 0;
 
         int transformLength = object.containsKey(TRANSFORM)
-            ? object.getJsonObject(TRANSFORM).getInt(LENGTH)
+            ? object.getJsonObject(TRANSFORM).getInt(LENGTH, -1)
             : -1;
+
+        List<Long> transformAuthorizations = null;
+        if (object.containsKey(TRANSFORM_AUTHORIZATIONS))
+        {
+            transformAuthorizations = new LinkedList<>();
+            for (JsonValue item : object.getJsonArray(TRANSFORM_AUTHORIZATIONS))
+            {
+                transformAuthorizations.add(((JsonNumber) item).longValue());
+            }
+        }
 
         boolean read = object.containsKey(CAPABILITY)
             ? object.getString(CAPABILITY).equals(READ)
@@ -99,6 +111,6 @@ public class TestModelConfigAdapter extends ConfigAdapter<ModelConfig, JsonValue
 
         ValidateConfig validateConfig = validate.adaptFromJsonObject(object);
 
-        return new TestModelConfig(length, catalogs, read, transformLength, fields, validateConfig);
+        return new TestModelConfig(length, catalogs, read, transformLength, fields, validateConfig, transformAuthorizations);
     }
 }